treewide: Move environment.go into a separate package
File environment.go was completely self-contained and did not rely on any
other functions from the "ovn" package. It is therefore moved into its own
package to allow it to be more easily importable from other packages.

(cherry picked from commit f28a8e563ff02b435f16030b44b964eb18b44d59)
