+++ microovn/ovn/environment/environment.go
@@ -1,4 +1,4 @@
-package ovn
+package environment
 
 import (
 	"context"
@@ -31,9 +31,9 @@ OVN_SB_CONNECT="{{ .sbConnect }}"
 OVN_LOCAL_IP="{{ .localAddr }}"
 `))
 
-// networkProtocol returns appropriate network protocol that should be used
+// NetworkProtocol returns appropriate network protocol that should be used
 // by OVN services.
-func networkProtocol(ctx context.Context, s state.State) string {
+func NetworkProtocol(ctx context.Context, s state.State) string {
 	_, _, err := certificates.GetCA(ctx, s)
 	if err != nil {
 		return "tcp"
@@ -41,8 +41,11 @@ func networkProtocol(ctx context.Context, s state.State) string {
 	return "ssl"
 }
 
-// Builds environment variable strings for OVN.
-func environmentString(ctx context.Context, s state.State, port int) (string, string, error) {
+// ConnectionString builds a string that defines connection endpoints of OVN
+// cluster services. It can be used to tell services (i.e. ovn-controller) about
+// the location of OVN central services.
+// Example return value: "ssl:10.0.0.1:6641,ssl:10.0.0.2:6641,ssl:10.0.0.3:6641"
+func ConnectionString(ctx context.Context, s state.State, port int) (string, string, error) {
 	var err error
 	var servers []database.Service
 	var clusterMap map[string]cluster.CoreClusterMember
@@ -71,7 +74,7 @@ func environmentString(ctx context.Context, s state.State, port int) (string, st
 
 	addresses := make([]string, 0, len(servers))
 	var initialString string
-	protocol := networkProtocol(ctx, s)
+	protocol := NetworkProtocol(ctx, s)
 	for i, server := range servers {
 		member := clusterMap[server.Member]
 		memberAddr, err := netip.ParseAddrPort(member.Address)
@@ -98,14 +101,15 @@ func environmentString(ctx context.Context, s state.State, port int) (string, st
 	return strings.Join(addresses, ","), initialString, nil
 }
 
-func generateEnvironment(ctx context.Context, s state.State) error {
+// GenerateEnvironment generates the OVN environment file.
+func GenerateEnvironment(ctx context.Context, s state.State) error {
 	// Get the servers.
-	nbConnect, nbInitial, err := environmentString(ctx, s, 6641)
+	nbConnect, nbInitial, err := ConnectionString(ctx, s, 6641)
 	if err != nil {
 		return err
 	}
 
-	sbConnect, sbInitial, err := environmentString(ctx, s, 6642)
+	sbConnect, sbInitial, err := ConnectionString(ctx, s, 6642)
 	if err != nil {
 		return err
 	}
@@ -152,7 +156,8 @@ func generateEnvironment(ctx context.Context, s state.State) error {
 	return nil
 }
 
-func createPaths() error {
+// CreatePaths creates the required directories for OVN.
+func CreatePaths() error {
 	// Create our various paths.
 	for _, path := range paths.RequiredDirs() {
 		err := os.MkdirAll(path, 0700)
@@ -164,9 +169,9 @@ func createPaths() error {
 	return nil
 }
 
-// cleanupPaths backs up directories defined by paths.BackupDirs and then removes directories
-// created by createPaths function. This effectively removes any data created during MicroOVN runtime.
-func cleanupPaths() error {
+// CleanupPaths backs up directories defined by paths.BackupDirs and then removes directories
+// created by CreatePaths function. This effectively removes any data created during MicroOVN runtime.
+func CleanupPaths() error {
 	var errs []error
 
 	// Create timestamped backup dir
