+++ microovn/ovn/start.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/canonical/microovn/microovn/node"
 	ovnCmd "github.com/canonical/microovn/microovn/ovn/cmd"
+	"github.com/canonical/microovn/microovn/ovn/environment"
 	"github.com/canonical/microovn/microovn/ovn/ovsdb"
 )
 
@@ -23,13 +24,13 @@ func Start(ctx context.Context, s state.State) error {
 	}
 
 	// Make sure the storage exists.
-	err = createPaths()
+	err = environment.CreatePaths()
 	if err != nil {
 		return err
 	}
 
 	// Re-generate the configuration.
-	err = generateEnvironment(ctx, s)
+	err = environment.GenerateEnvironment(ctx, s)
 	if err != nil {
 		return fmt.Errorf("failed to generate the daemon configuration: %w", err)
 	}
@@ -46,7 +47,7 @@ func Start(ctx context.Context, s state.State) error {
 		}
 	}
 	// Reconfigure OVS to use OVN.
-	sbConnect, _, err := environmentString(ctx, s, 6642)
+	sbConnect, _, err := environment.ConnectionString(ctx, s, 6642)
 	if err != nil {
 		return fmt.Errorf("failed to get OVN SB connect string: %w", err)
 	}
